Make effect stoppers & assistance/offense immunity validations work only with targeted effects

+++ core/src/sol/svc/vast/vaste_maintain/vaste_effect_projected_unprojected.rs
@@ -1,5 +1,5 @@
 use crate::{
-    ad,
+    ac, ad,
     sol::{
         ItemKey,
         svc::{EffectSpec, vast::Vast},
@@ -16,33 +16,35 @@ impl Vast {
         projectee_item_key: ItemKey,
         projectee_item: &UadItem,
     ) {
-        if !a_effect.stop_ids.is_empty()
-            && let Some(projectee_fit_key) = projectee_item.get_fit_key()
-        {
-            let projectee_fit_data = self.fit_datas.get_mut(&projectee_fit_key).unwrap();
-            let stopper = EffectSpec::new(projector_item_key, a_effect.id);
-            for stop_a_effect_id in a_effect.stop_ids.iter() {
-                let stopped = EffectSpec::new(projectee_item_key, *stop_a_effect_id);
-                projectee_fit_data.stopped_effects.add_entry(stopped, stopper);
+        if a_effect.category == ac::effcats::TARGET {
+            if !a_effect.stop_ids.is_empty()
+                && let Some(projectee_fit_key) = projectee_item.get_fit_key()
+            {
+                let projectee_fit_data = self.fit_datas.get_mut(&projectee_fit_key).unwrap();
+                let stopper = EffectSpec::new(projector_item_key, a_effect.id);
+                for stop_a_effect_id in a_effect.stop_ids.iter() {
+                    let stopped = EffectSpec::new(projectee_item_key, *stop_a_effect_id);
+                    projectee_fit_data.stopped_effects.add_entry(stopped, stopper);
+                }
             }
-        }
-        if a_effect.is_assist
-            && let Some(projector_fit_key) = projector_item.get_fit_key()
-        {
-            let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
-            let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
-            projector_fit_data
-                .blockable_assistance
-                .add_entry(projector_espec, projectee_item_key);
-        }
-        if is_offense_blockable(projector_item, a_effect)
-            && let Some(projector_fit_key) = projector_item.get_fit_key()
-        {
-            let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
-            let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
-            projector_fit_data
-                .blockable_offense
-                .add_entry(projector_espec, projectee_item_key);
+            if a_effect.is_assist
+                && let Some(projector_fit_key) = projector_item.get_fit_key()
+            {
+                let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
+                let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
+                projector_fit_data
+                    .blockable_assistance
+                    .add_entry(projector_espec, projectee_item_key);
+            }
+            if is_offense_blockable(projector_item, a_effect)
+                && let Some(projector_fit_key) = projector_item.get_fit_key()
+            {
+                let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
+                let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
+                projector_fit_data
+                    .blockable_offense
+                    .add_entry(projector_espec, projectee_item_key);
+            }
         }
     }
     pub(in crate::sol::svc) fn effect_unprojected(
@@ -53,33 +55,35 @@ impl Vast {
         projectee_item_key: ItemKey,
         projectee_item: &UadItem,
     ) {
-        if !a_effect.stop_ids.is_empty()
-            && let Some(projectee_fit_key) = projectee_item.get_fit_key()
-        {
-            let projectee_fit_data = self.fit_datas.get_mut(&projectee_fit_key).unwrap();
-            let stopper = EffectSpec::new(projector_item_key, a_effect.id);
-            for stop_a_effect_id in a_effect.stop_ids.iter() {
-                let stopped = EffectSpec::new(projectee_item_key, *stop_a_effect_id);
-                projectee_fit_data.stopped_effects.remove_entry(&stopped, &stopper);
+        if a_effect.category == ac::effcats::TARGET {
+            if !a_effect.stop_ids.is_empty()
+                && let Some(projectee_fit_key) = projectee_item.get_fit_key()
+            {
+                let projectee_fit_data = self.fit_datas.get_mut(&projectee_fit_key).unwrap();
+                let stopper = EffectSpec::new(projector_item_key, a_effect.id);
+                for stop_a_effect_id in a_effect.stop_ids.iter() {
+                    let stopped = EffectSpec::new(projectee_item_key, *stop_a_effect_id);
+                    projectee_fit_data.stopped_effects.remove_entry(&stopped, &stopper);
+                }
             }
-        }
-        if a_effect.is_assist
-            && let Some(projector_fit_key) = projector_item.get_fit_key()
-        {
-            let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
-            let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
-            projector_fit_data
-                .blockable_assistance
-                .remove_entry(&projector_espec, &projectee_item_key);
-        }
-        if is_offense_blockable(projector_item, a_effect)
-            && let Some(projector_fit_key) = projector_item.get_fit_key()
-        {
-            let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
-            let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
-            projector_fit_data
-                .blockable_offense
-                .remove_entry(&projector_espec, &projectee_item_key);
+            if a_effect.is_assist
+                && let Some(projector_fit_key) = projector_item.get_fit_key()
+            {
+                let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
+                let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
+                projector_fit_data
+                    .blockable_assistance
+                    .remove_entry(&projector_espec, &projectee_item_key);
+            }
+            if is_offense_blockable(projector_item, a_effect)
+                && let Some(projector_fit_key) = projector_item.get_fit_key()
+            {
+                let projector_fit_data = self.fit_datas.get_mut(&projector_fit_key).unwrap();
+                let projector_espec = EffectSpec::new(projector_item_key, a_effect.id);
+                projector_fit_data
+                    .blockable_offense
+                    .remove_entry(&projector_espec, &projectee_item_key);
+            }
         }
     }
 }
